+++ ay-3-8910/notes.h
@@ -1,2 +1,2 @@
-double note_to_freq(char note, int flat, int sharp, int octave, int sub);
+double note_to_freq(char note, int flat, int sharp, int octave, double sub);
 int note_to_ed(char note, int flat, int sharp, int octave);
